Set Postgres client encoding to enable RedShift connections
Pgcli is now using psycopg3 under the hood. This throws an error
when it tries to talk to RedShift.

psycopg.NotSupportedError: codec not available in Python: b'UNICODE'

It's something to do how RedShift is reporting its encoding or
something like that. I'm not entirely sure, but setting Postgres'
client encoding resolves the issue.