Adding prettier formatting
I'm adding support for prettier formatting using null-ls formatters.
The typescript-language-server integration in Vim (not sure if that's
called tsserver) doesn't appear to provide formatting.

The prettier website recommends getting an extra plugin. I tried using
Neoformat (https://prettier.io/docs/en/vim#neoformathttpsgithubcomsbdchdneoformat)
but then I realized that null-ls actually was able to get me formatting
for free. So I opted to remove the Neoformat setup and stuck with null-ls
instead.

For some strange reason, this is no longer working though, and I'm
not sure why. I did recently modify my eslintrc to make it simpler
but I doubt that could be the culprit.

So currently, the code doesn't format on save, or when I was "lf"
with which key bindings.