Enable eslint formatting
Null-ls supports both eslint and eslint_d. eslint_d is basically
a faster version of eslint. It cuts down the formatting time from
100's of milliseconds to less than 200ms.

I'm not entirely sure why the configuration is like that, however,
eslint runs with --fix-dryrun which fixes the changes but doesn't
make any filesystem changes. There's also a format that outputs this
to JSON. So it looks like the intention is to process the output.
I saw somewhere that it's too slow to be used for formatting.
See the args that eslint run with below:
https://github.com/jose-elias-alvarez/null-ls.nvim/blob/0010ea927ab7c09ef0ce9bf28c2b573fc302f5a7/lua/null-ls/builtins/formatting/eslint.lua#L28

Eslint_d on the other hand, does have a --fix command. I had to do
a couple restarts of Lvim, but once I did that, Eslint formatted the
code.