Add the ability to fold levels using <number>zf
I wanted to be able to quickly collapse class methods, or entire
classes. This is controlled in Vim by the fold level. When the
fold level is really high, there is no code folding.

When the fold level is set to zero, everything at the first
indent level or tab position gets folded. Level 1 folds
the next level. In Python for example, 0 would fold all
classes and 1 would fold the methods of all classes.

I can now do this using <number>zf